+++ CanvasPlusPlayground/Features/Pinned Items/PinnedItemsView.swift
@@ -18,15 +18,22 @@ struct PinnedItemsView: View {
     var body: some View {
         List(sortedTypes, id: \.self) { type in
             Section {
+                let items = pinnedItemsManager.pinnedItemsByType[type] ?? []
                 ScrollView(.horizontal, showsIndicators: false) {
-                    HStack {
-                        ForEach(pinnedItemsManager.pinnedItemsByType[type] ?? []) { item in
+                    LazyHGrid(
+                        rows: .init(
+                            repeating: .init(.flexible(minimum: 100, maximum: 400)),
+                            count: min(3, items.count)
+                        )
+                    ) {
+                        ForEach(items) { item in
                             PinnedItemCard(item: item)
                         }
                     }
                 }
                 .scrollTargetBehavior(.paging)
                 .listRowSeparator(.hidden)
+                .fixedSize(horizontal: false, vertical: true)
 
             } header: {
                 Text(type.displayName)
@@ -37,65 +44,9 @@ struct PinnedItemsView: View {
             }
 
         }
+        .navigationTitle("Pinned")
         #if os(macOS)
         .navigationSplitViewColumnWidth(min: 350, ideal: 400)
         #endif
-    }
-}
-
-struct PinnedItemCard: View {
-    let item: PinnedItem
-
-    var body: some View {
-        AsyncView {
-            await item.itemData()
-        } content: { itemData in
-            switch itemData.modelData {
-            case .announcement(let announcement):
-                PinnedAnnouncementCard(
-                    announcement: announcement,
-                    course: itemData.course
-                )
-            default: Text("Got \(itemData)")
-            }
-        } placeholder: {
-            Text("Loading...")
-        }
-    }
-}
-
-struct PinnedAnnouncementCard: View {
-    let announcement: Announcement
-    let course: Course
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            Text(course.displayName.uppercased())
-                .font(.caption)
-                .foregroundStyle(course.rgbColors?.color ?? .accentColor)
-
-            VStack(alignment: .leading, spacing: 3) {
-                Text(announcement.title ?? "")
-                    .font(.headline)
-                    .fontDesign(.rounded)
-                    .bold()
-
-                Text(
-                    announcement.message?
-                        .stripHTML()
-                        .trimmingCharacters(
-                            in: .whitespacesAndNewlines
-                        )
-                    ?? ""
-                )
-                .lineLimit(2)
-            }
-        }
-        .frame(width: 250)
-        .padding(12)
-        .background {
-            RoundedRectangle(cornerRadius: 16.0)
-                .fill(.secondary.opacity(0.3))
-        }
     }
 }
